refactor(ui): split theme.ts (907 → 6 files) by domain
Ninth worst-first file-size refactor; first TSX target. The design-token
file is split into per-domain modules, with theme.ts kept as a 167-line
barrel that re-exports everything plus the cn/buttonClass/inputClass/
cardClass/badgeClass/modalClass utility helpers.

  theme.ts             167  barrel + utility class-name composers
  themeSpacing.ts      182  margin/padding/gap/stack tokens
  themeTypography.ts    63  heading/body/size/weight/family tokens
  themeComponents.ts   155  button, input, card, badge, toast, alert, modal, section
  themeColors.ts       210  status, severity, timing, category, moduleColor, brand, gauge
  themeLayout.ts       139  sizing, icon, radius, border, flex/grid patterns

All files under the 800-line TSX warn budget; only theme.ts is anywhere
near it after the split. No callers updated: the ~90 existing
`import { spacing, button, cn } from '.../styles/theme'` sites keep
working as-is.

Drive-by: a single biome-ignore noBarrelFile suppressor on theme.ts
documenting why the re-exports are intentional for backwards compat.

Vite build succeeds; biome check clean across all six files.
